Better structure for future expansion :)
