+++ include/mu-emu/core.h
