+++ include/mu-emu/rv32i.h
@@ -1,6 +1,5 @@
 #pragma once
-#include "../vm.h"
-#include "inst.h"
+#include "mu-emu/core.h"
 
 namespace MuEmu { namespace ISA { namespace RV32I {
     // X for unknown
@@ -12,6 +11,26 @@ namespace MuEmu { namespace ISA { namespace RV32I {
         public:
             ExecutionStrategy(MuEmu::VM &vm) : MuEmu::VM::ExecutionStrategy(vm) {}
             void Execute() override;
+    };
+
+    class Instruction {
+        private:
+            int32_t word;
+        public:
+            Instruction(int32_t word) : word(word) {}
+            uint8_t Rd();
+            uint8_t Rs1();
+            uint8_t Rs2();
+            int8_t Opcode();
+            int8_t Func3();
+            int8_t Func7();
+            int32_t ImmI();
+            int32_t ImmS();
+            int32_t ImmB();
+            int32_t ImmU();
+            int32_t ImmJ();
+            InstructionType Type();
+            std::string Mnemonic();
     };
 
     class InstructionSetEntry {
