+++ src/rv32i/rv32i.cpp
@@ -1,4 +1,4 @@
-#include "rv32i.h"
+#include "mu-emu/rv32i.h"
 
 using namespace MuEmu::ISA::RV32I;
